Make calibration readings double because they are now in radians
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@5049 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ mechanism/hardware_interface/include/hardware_interface/hardware_interface.h
@@ -63,8 +63,8 @@ public:
   double velocity_;
 
   bool calibration_reading_;
-  int32_t last_calibration_high_transition_;  // Last transition from high to low.
-  int32_t last_calibration_low_transition_;
+  double last_calibration_high_transition_;  // Last transition from high to low.
+  double last_calibration_low_transition_;
 
   bool is_enabled_;
   bool run_stop_hit_;
